RavenDB-4842 renamed Smuggler files
