+++ src/Raven.Server/Documents/PeriodicExport/IncrementalExport.cs
@@ -1,6 +1,7 @@
 using System.IO;
 using Raven.Server.ServerWide.Context;
 using Raven.Server.Smuggler;
+using Raven.Server.Smuggler.Documents;
 
 namespace Raven.Server.Documents.PeriodicExport
 {
@@ -8,7 +9,7 @@ namespace Raven.Server.Documents.PeriodicExport
     {
         private const string IncrementalExportStateFile = "IncrementalExport.state.json";
 
-        public static void ReadLastEtagsFromFile(string exportDirectory, DocumentsOperationContext context, DatabaseDataExporter dataExporter)
+        public static void ReadLastEtagsFromFile(string exportDirectory, DocumentsOperationContext context, SmugglerExporter smugglerExporter)
         {
             var etagFileLocation = Path.Combine(exportDirectory, IncrementalExportStateFile);
             if (File.Exists(etagFileLocation) == false)
@@ -17,7 +18,7 @@ namespace Raven.Server.Documents.PeriodicExport
             using (var fileStream = new FileStream(etagFileLocation, FileMode.Open))
             {
                 var reader = context.ReadForMemory(fileStream, IncrementalExportStateFile);
-                reader.TryGet("LastDocsEtag", out dataExporter.StartDocsEtag);
+                reader.TryGet("LastDocsEtag", out smugglerExporter.StartDocsEtag);
             }
         }
     }
