+++ src/Raven.Server/Json/JsonDeserializationServer.cs
@@ -9,6 +9,8 @@ using Raven.Server.Documents.PeriodicExport;
 using Raven.Server.Documents.SqlReplication;
 using Raven.Server.Documents.Versioning;
 using Raven.Server.Smuggler;
+using Raven.Server.Smuggler.Documents;
+using Raven.Server.Smuggler.Documents.Data;
 using Sparrow.Json;
 
 namespace Raven.Server.Json
