+++ src/Raven.Server/Smuggler/Documents/Data/ExportResult.cs
@@ -1,4 +1,4 @@
-namespace Raven.Server.Smuggler
+namespace Raven.Server.Smuggler.Documents.Data
 {
     public class ExportResult
     {
